fix: Move utility functions to utils package (#1012)
