+++ src/internal/model_render_test.go
@@ -1,7 +1,6 @@
 package internal
 
 import (
-	"fmt"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -11,6 +10,8 @@ import (
 	"github.com/charmbracelet/x/exp/term/ansi"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/yorukot/superfile/src/internal/utils"
 )
 
 func TestFilePreviewRenderWithDimensions(t *testing.T) {
@@ -28,7 +29,7 @@ func TestFilePreviewRenderWithDimensions(t *testing.T) {
 	curTestDir := filepath.Join(testDir, "TestFilePreviewRender")
 
 	// Cleanup is taken care by TestMain()
-	setupDirectories(t, curTestDir)
+	utils.SetupDirectories(t, curTestDir)
 
 	testdata := []struct {
 		name            string
@@ -135,7 +136,7 @@ func TestFilePreviewRenderWithDimensions(t *testing.T) {
 	for i, tt := range testdata {
 		t.Run(tt.name, func(t *testing.T) {
 			curDir := filepath.Join(curTestDir, "dir"+strconv.Itoa(i))
-			setupDirectories(t, curDir)
+			utils.SetupDirectories(t, curDir)
 			filePath := filepath.Join(curDir, tt.fileName)
 			err := os.WriteFile(filePath, []byte(tt.fileContent), 0644)
 			require.NoError(t, err)
@@ -176,69 +177,4 @@ func normalizeOutput(output string) string {
 
 	// Join the lines back together
 	return strings.Join(filteredLines, "\n")
-}
-
-func TestReadFileContent(t *testing.T) {
-	curTestDir := filepath.Join(testDir, "TestReadFileContent")
-	setupDirectories(t, curTestDir)
-
-	testdata := []struct {
-		name          string
-		content       []byte
-		maxLineLength int
-		previewLine   int
-		expected      string
-	}{
-		{
-			name:          "regular UTF-8 file",
-			content:       []byte("line1\nline2\nline3"),
-			maxLineLength: 100,
-			previewLine:   5,
-			expected:      "line1\nline2\nline3\n",
-		},
-		{
-			name:          "UTF-8 BOM file",
-			content:       []byte("\xEF\xBB\xBFline1\nline2\nline3"),
-			maxLineLength: 100,
-			previewLine:   5,
-			expected:      "line1\nline2\nline3\n",
-		},
-		{
-			name:          "limited preview lines",
-			content:       []byte("line1\nline2\nline3\nline4"),
-			maxLineLength: 100,
-			previewLine:   2,
-			expected:      "line1\nline2\n",
-		},
-	}
-
-	for i, tt := range testdata {
-		t.Run(tt.name, func(t *testing.T) {
-			testFile := filepath.Join(curTestDir, fmt.Sprintf("test_file_%d.txt", i))
-			setupFilesWithData(t, tt.content, testFile)
-
-			result, err := readFileContent(testFile, tt.maxLineLength, tt.previewLine)
-			require.NoError(t, err)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
-
-func TestReadFileContentBOMHandling(t *testing.T) {
-	curTestDir := filepath.Join(testDir, "TestBOMHandling")
-	setupDirectories(t, curTestDir)
-
-	// Write a file prefixed with UTF-8 BOM
-	bomContent := []byte("\xEF\xBB\xBFHello, World!\nSecond line")
-	bomFile := filepath.Join(curTestDir, "bom_file.txt")
-	setupFilesWithData(t, bomContent, bomFile)
-
-	result, err := readFileContent(bomFile, 100, 10)
-	require.NoError(t, err)
-
-	// Verify BOM is removed and content is correct
-	assert.True(t, strings.HasPrefix(result, "Hello, World!"),
-		"Content should start with expected text, got: %q", result)
-	assert.NotContains(t, result, "\uFEFF",
-		"BOM character should be removed from output: %q", result)
 }
