+++ src/internal/utils/test_utils.go
@@ -1,5 +1,14 @@
 package utils
 
+import (
+	"os"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+var SampleDataBytes = []byte("This is sample") //nolint: gochecknoglobals // Effectively const
+
 type TestTOMLType struct {
 	SampleBool  bool     `toml:"sample_bool"`
 	SampleInt   int      `toml:"sample_int"`
@@ -22,4 +31,27 @@ func (t TestTOMLMissingIgnorerType) GetIgnoreMissingFields() bool {
 func (t TestTOMLMissingIgnorerType) WithIgnoreMissing(val bool) TestTOMLMissingIgnorerType {
 	t.IgnoreMissing = val
 	return t
+}
+
+func SetupDirectories(t *testing.T, dirs ...string) {
+	t.Helper()
+	for _, dir := range dirs {
+		if dir == "" {
+			continue
+		}
+		err := os.MkdirAll(dir, 0755)
+		require.NoError(t, err)
+	}
+}
+
+func SetupFilesWithData(t *testing.T, data []byte, files ...string) {
+	t.Helper()
+	for _, file := range files {
+		err := os.WriteFile(file, data, 0644)
+		require.NoError(t, err)
+	}
+}
+
+func SetupFiles(t *testing.T, files ...string) {
+	SetupFilesWithData(t, SampleDataBytes, files...)
 }
